fix backup

+++ ubuntu/common/files/usr/local/bin/sync-to-nas.sh
@@ -2,15 +2,16 @@
 
 set -e
 
-source "$( dirname "${BASH_SOURCE[0]}" )/../functions.sh"
+source "$( dirname "${BASH_SOURCE[0]}" )/functions.sh"
 
 for bak in /boot/bak*
 do
+    [ -d "${bak}" ] || continue
     rsync -arPx --delete /boot/ "${bak}"
 done
 rsync -arPx --delete /boot/ root@nas:/mnt/e/$(hostname)/boot
 
-DATASETS=($(zfs list -o name))
+DATASETS=($(zfs list -o name | grep -v root/docker))
 
 for x in "${DATASETS[@]}"
 do
